在UserWithdraw，User，Task，HunterTask中新增审核对象人编号字段

+++ tc-company/src/main/java/com/tc/controller/MessageController.java
@@ -2,28 +2,36 @@ package com.tc.controller;
 
 import com.google.gson.Gson;
 import com.tc.db.entity.Message;
+import com.tc.db.entity.User;
 import com.tc.db.enums.MessageState;
+import com.tc.db.enums.UserCategory;
+import com.tc.db.enums.UserState;
 import com.tc.dto.LongIds;
 import com.tc.dto.Result;
 import com.tc.dto.message.AddMessage;
 import com.tc.dto.message.ModifyMessage;
 import com.tc.dto.message.MyCondition;
 import com.tc.dto.message.QueryMessage;
+import com.tc.dto.trans.TransStates;
 import com.tc.dto.user.QueryUser;
 import com.tc.exception.DBException;
 import com.tc.exception.ValidException;
+import com.tc.service.AdminService;
 import com.tc.service.MessageService;
 import com.tc.service.UserService;
 import com.tc.until.ListUtils;
+import com.tc.until.PageRequest;
 import com.tc.until.StringResourceCenter;
 import io.swagger.annotations.ApiOperation;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import java.util.Arrays;
 import java.util.List;
@@ -53,6 +61,8 @@ public class MessageController {
     @Autowired
     private UserService userService;
 
+    @Autowired
+    private AdminService adminService;
 
     /**
      * 根据查询条件，获取消息列表
@@ -61,9 +71,11 @@ public class MessageController {
      */
     @PostMapping("/query")
     @ApiOperation(value = "根据查询条件，获取消息列表")
-    public Result all(@RequestBody QueryMessage queryMessage){
+    public Result all(HttpServletRequest request, @RequestBody QueryMessage queryMessage){
+        Long me = Long.parseLong(request.getAttribute(StringResourceCenter.USER_ID).toString());
+        queryMessage.setSort(new Sort(new Sort.Order(Sort.Direction.DESC,Message.STATE),new Sort.Order(Sort.Direction.DESC,Message.CREATE_TIME)));
         Page<Message> query = messageService.findByQuery(queryMessage);
-        return Result.init(Message.toListByIndex(query.getContent()),queryMessage);
+        return Result.init(Message.toListByIndex(query.getContent(),me),queryMessage.append(query.getTotalElements(),(long)query.getTotalPages()).clearSort());
     }
 
     /**
@@ -86,11 +98,13 @@ public class MessageController {
      */
     @PostMapping
     @ApiOperation(value = "添加新消息")
-    public Message add(@Valid @RequestBody AddMessage addMessage, BindingResult bindingResult){
+    public Message add(HttpServletRequest request, @Valid @RequestBody AddMessage addMessage, BindingResult bindingResult){
+        Long me = Long.parseLong(request.getAttribute(StringResourceCenter.USER_ID).toString());
         if (bindingResult.hasErrors()){
             throw new ValidException(bindingResult.getFieldErrors());
         }
-        Message message = messageService.save(AddMessage.toMessage(addMessage));
+        Message message = messageService.save(AddMessage.toMessage(addMessage,me));
+        message.setCreation(adminService.findOne(me));
         return Message.toDetail(message);
     }
 
@@ -102,13 +116,17 @@ public class MessageController {
      */
     @PutMapping
     @ApiOperation(value = "修改消息")
-    public Message update(@Valid @RequestBody ModifyMessage modifyMessage, BindingResult bindingResult){
+    public Message update(HttpServletRequest request, @Valid @RequestBody ModifyMessage modifyMessage, BindingResult bindingResult){
+        Long me = Long.parseLong(request.getAttribute(StringResourceCenter.USER_ID).toString());
         if (bindingResult.hasErrors()){
             throw new ValidException(bindingResult.getFieldErrors());
         }
         Message old = messageService.findOne(modifyMessage.getId());
         if (old == null){
             throw new DBException(StringResourceCenter.DB_QUERY_FAILED);
+        }
+        if (!old.getCreationId().equals(me)){
+            throw new ValidException(StringResourceCenter.VALIDATOR_AUTHORITY_FAILED);
         }
         if (!ModifyMessage.isUpdate(old,modifyMessage)){
             throw new ValidException(StringResourceCenter.VALIDATOR_UPDATE_ABNORMAL);
@@ -123,10 +141,14 @@ public class MessageController {
      */
     @GetMapping("/issue/{id:\\d+}")
     @ApiOperation(value = "管理员发布消息")
-    public void issueMessage(@PathVariable("id") Long id){
+    public void issueMessage(HttpServletRequest request, @PathVariable("id") Long id){
+        Long me = Long.parseLong(request.getAttribute(StringResourceCenter.USER_ID).toString());
         Message query = messageService.findOne(id);
         if (query == null){
             throw new DBException(StringResourceCenter.DB_QUERY_FAILED);
+        }
+        if (!query.getCreationId().equals(me)){
+            throw new ValidException(StringResourceCenter.VALIDATOR_AUTHORITY_FAILED);
         }
         if (StringUtils.isEmpty(query.getLookCondition()) || query.getType() == null){
             throw new ValidException("请设置查看条件后，在发布消息");
@@ -143,13 +165,17 @@ public class MessageController {
      */
     @GetMapping("/out/{id:\\d+}")
     @ApiOperation(value = "管理员撤回消息")
-    public void outMessage(@PathVariable("id") Long id){
+    public void outMessage(HttpServletRequest request, @PathVariable("id") Long id){
+        Long me = Long.parseLong(request.getAttribute(StringResourceCenter.USER_ID).toString());
         Message query = messageService.findOne(id);
         if (query == null){
             throw new DBException(StringResourceCenter.DB_QUERY_FAILED);
         }
         if (StringUtils.isEmpty(query.getLookCondition()) || query.getType() == null){
             throw new ValidException("请设置查看条件后，在发布消息");
+        }
+        if (!query.getCreationId().equals(me)){
+            throw new ValidException(StringResourceCenter.VALIDATOR_AUTHORITY_FAILED);
         }
         boolean isSuccess = messageService.updateState(id,MessageState.STOP);
         if (!isSuccess){
@@ -163,7 +189,15 @@ public class MessageController {
      */
     @DeleteMapping("/{id:\\d+}")
     @ApiOperation(value = "删除单个消息")
-    public void delete(@PathVariable("id") Long id){
+    public void delete(HttpServletRequest request, @PathVariable("id") Long id){
+        Long me = Long.parseLong(request.getAttribute(StringResourceCenter.USER_ID).toString());
+        Message query = messageService.findOne(id);
+        if (query == null){
+            throw new DBException(StringResourceCenter.DB_QUERY_FAILED);
+        }
+        if (!query.getCreationId().equals(me)){
+            throw new ValidException(StringResourceCenter.VALIDATOR_AUTHORITY_FAILED);
+        }
         boolean isSuccess = messageService.deleteById(id);
         if (!isSuccess){
             throw new DBException(StringResourceCenter.DB_DELETE_ABNORMAL);
@@ -177,9 +211,18 @@ public class MessageController {
      */
     @PutMapping("/condition")
     @ApiOperation(value = "设置消息可查看人的条件")
-    public Message addCondition(@Valid @RequestBody MyCondition condition, BindingResult bindingResult){
+    public Message addCondition(HttpServletRequest request,@Valid @RequestBody MyCondition condition, BindingResult bindingResult){
+        Long me = Long.parseLong(request.getAttribute(StringResourceCenter.USER_ID).toString());
         if (bindingResult.hasErrors()){
             throw new ValidException(bindingResult.getFieldErrors());
+        }
+
+        Message message = messageService.findOne(condition.getId());
+        if (message == null){
+            throw new DBException(StringResourceCenter.DB_QUERY_FAILED);
+        }
+        if (!message.getCreationId().equals(me)){
+            throw new ValidException(StringResourceCenter.VALIDATOR_AUTHORITY_FAILED);
         }
         String jsonContext = null;
         long count = 0;
@@ -219,10 +262,6 @@ public class MessageController {
             throw new ValidException("没有要添加的条件");
         }
         condition.setContext(jsonContext);
-        Message message = messageService.findOne(condition.getId());
-        if (message == null){
-            throw new DBException(StringResourceCenter.DB_QUERY_FAILED);
-        }
         Message result = messageService.save(MyCondition.toMessage(message,condition));
         if (result == null){
             throw new DBException(StringResourceCenter.DB_UPDATE_ABNORMAL);
@@ -231,6 +270,71 @@ public class MessageController {
         return Message.toDetail(result);
     }
 
+
+    /**
+     * 根据查询条件获取用户列表
+     *
+     * @param queryUser 用户查询条件
+     * @return
+     */
+    @PostMapping("/condition/user")
+    @ApiOperation(value = "根据查询条件获取创建消息时需要的用户列表")
+    public Result all(@RequestBody QueryUser queryUser) {
+
+        if (queryUser.getCategory() == null){
+            queryUser.setCategories(UserCategory.byQueryUsers());
+        }
+
+        Page<User> queryUsers = userService.findByQueryUser(queryUser);
+        return Result.init(User.toIndexAsList(queryUsers.getContent()), queryUser.append(queryUsers.getTotalElements(), (long) queryUsers.getTotalPages()));
+    }
+
+
+    /**
+     * 获取可查看指定消息的用户列表
+     * @param id
+     * @return
+     */
+    @PostMapping("/condition/user/{id:\\d+}")
+    @ApiOperation(value = "获取可查看指定消息的用户列表")
+    public Result conditionUser(@PathVariable("id") Long id,@RequestBody PageRequest page){
+        Message result = messageService.findOne(id);
+        if (result == null){
+            throw new DBException(StringResourceCenter.DB_QUERY_FAILED);
+        }
+        if (!StringUtils.isEmpty(result.getLookCondition()) && result.getType() != null){
+            switch (result.getType()){
+                case CONDITION:
+                    QueryUser queryUser;
+                    try {
+                        queryUser = new Gson().fromJson(result.getLookCondition(),QueryUser.class);
+                    }catch (Exception e){
+                        throw new DBException("JSON转换异常");
+                    }
+                    queryUser.setPageSize(page.getPageSize());
+                    queryUser.setPageNumber(page.getPageNumber());
+                    Page<User> queryUsers = userService.findByQueryUser(queryUser);
+                    return Result.init(User.toIndexAsList(queryUsers.getContent()),page.append(queryUsers.getTotalElements(),(long)queryUsers.getTotalPages()));
+                case APPOINT:
+                    LongIds longIds;
+                    try {
+                        longIds = new Gson().fromJson(result.getLookCondition(),LongIds.class);
+                    }catch (Exception e){
+                        throw new DBException("JSON转换异常");
+                    }
+                    Page<User> idsUsers = userService.findByIdsAndPage(longIds.getIds(),page);
+                    return Result.init(User.toIndexAsList(idsUsers.getContent()),page.append(idsUsers.getTotalElements(),(long)idsUsers.getTotalPages()));
+                case ALL:
+                    throw new ValidException("所有人都可以看到该消息");
+                case IS_NULL:
+                    throw new ValidException("还未添加条件");
+                default:
+                    throw new ValidException("消息分类信息异常");
+            }
+        }else {
+            throw new DBException("没有添加任何条件");
+        }
+    }
 
     /**
      * 查看管理员设置的条件
@@ -279,7 +383,15 @@ public class MessageController {
      */
     @DeleteMapping("/condition/{id:\\d+}")
     @ApiOperation(value = "删除条件")
-    public void deleteCondition(@PathVariable("id") Long id){
+    public void deleteCondition(HttpServletRequest request,@PathVariable("id") Long id){
+        Long me = Long.parseLong(request.getAttribute(StringResourceCenter.USER_ID).toString());
+        Message message = messageService.findOne(id);
+        if (message == null){
+            throw new DBException(StringResourceCenter.DB_QUERY_FAILED);
+        }
+        if (!message.getCreationId().equals(me)){
+            throw new ValidException(StringResourceCenter.VALIDATOR_AUTHORITY_FAILED);
+        }
         boolean isSuccess = messageService.deleteCondition(id);
         if (!isSuccess){
             throw new DBException(StringResourceCenter.DB_DELETE_ABNORMAL);
